Updated Chrome & Firefox results after test fixes.
I also re-ran the tests in Firefox with the dom.messageChannel.enabled
option enabled. So that caused Firefox to pass a couple of more tests, but
no other UAs pass those tests, so it didn't make them drop off the
less-than-2 list. I left in the Chrome 39 and Firefox 34 results for now,
just for the sake of comparison. Anybody else should feel free to remove
them (since I added results for Chrome 41 and Firefox 36).