wai-aria: Combine aria-modal="true" subtests into a single test.
The various accessibility APIs handle exposure of aria-modal="true" quite
differently. For instance, in ATK, implementors must add STATE_MODAL to the
modal container. But in AX API, implementors must do nothing to the modal
container, and instead must completely remove all of the other elements from
the accessibility tree.

Differences such as these make it hard to test "implentability" across the
platforms using the current WPT ATTA framework which creates per-element
subtests. By preserving the existing results, but swapping the subtest names
for one of the platforms, we can generate results with wptreport which
accurately show the details of each PASS and FAIL as well as implementability
of the feature.