wai-aria: Combine results of two subtests for ATK into one subtest
There are two tests which have subtests with assertions only for ATK. The
reason for this is that ATK expects an additional accessibility event as
part of the implementation of the ARIA features in question, and our test
harness creates separate subtests for each element being examined. We can
preserve the detailed results for ATK without otherwise impacting the
cross-platform implementation report by combining the ATK-unique subtest
results with the previous subtest's results.