wai-aria: Filter out a test for something postponed until ARIA 1.2.
The ARIA spec does not have any normative statements regarding what user
agents should do if an author references a rendered element via the
aria-errormessage attribute with aria-invalid="false". The Working Group
concluded that if the ARIA spec needed to exit and re-enter CR, this
condition should be explicitly addressed in the spec. The ARIA spec did
not exit and re-enter CR, thus the associated test should not be part
of the implementation report for ARIA 1.1