wai-aria: Update results (#133)
* wai-aria: Updated results for AX API

Reflects removal of redundant and irrelevant assertions which had been
present in two tests. (The tests have since been updated.)

* wai-aria: Combine aria-modal="true" subtests into a single test.

The various accessibility APIs handle exposure of aria-modal="true" quite
differently. For instance, in ATK, implementors must add STATE_MODAL to the
modal container. But in AX API, implementors must do nothing to the modal
container, and instead must completely remove all of the other elements from
the accessibility tree.

Differences such as these make it hard to test "implentability" across the
platforms using the current WPT ATTA framework which creates per-element
subtests. By preserving the existing results, but swapping the subtest names
for one of the platforms, we can generate results with wptreport which
accurately show the details of each PASS and FAIL as well as implementability
of the feature.

* wai-aria: Combine results of two subtests for ATK into one subtest

There are two tests which have subtests with assertions only for ATK. The
reason for this is that ATK expects an additional accessibility event as
part of the implementation of the ARIA features in question, and our test
harness creates separate subtests for each element being examined. We can
preserve the detailed results for ATK without otherwise impacting the
cross-platform implementation report by combining the ATK-unique subtest
results with the previous subtest's results.

* wai-aria: Filter out a test for something postponed until ARIA 1.2.

The ARIA spec does not have any normative statements regarding what user
agents should do if an author references a rendered element via the
aria-errormessage attribute with aria-invalid="false". The Working Group
concluded that if the ARIA spec needed to exit and re-enter CR, this
condition should be explicitly addressed in the spec. The ARIA spec did
not exit and re-enter CR, thus the associated test should not be part
of the implementation report for ARIA 1.1

* wai-aria: Remove "placeholder" statement from README.md